refactor(cmd/web): remove duplicated error handlers and move to a separate file
